move passkey functionality to a separate feature
