+++ src/models.rs
@@ -3,9 +3,10 @@ use log::info;
 use crate::Error;
 
 pub mod db;
-pub mod keys;
 pub mod passwords;
 pub mod users;
+#[cfg(passkey)]
+pub mod keys;
 
 pub(crate) async fn init_db(client: &libsql_client::Client) -> Result<(), Error> {
     info!("initializing db");
